+++ service/globalaccelerator/doc.go
@@ -33,7 +33,7 @@
 // total of four addresses: two static IPv4 addresses and two static IPv6 addresses.
 // With a standard accelerator for IPv4, instead of using the addresses that
 // Global Accelerator provides, you can configure these entry points to be IPv4
-// addresses from your own IP address ranges that you bring toGlobal Accelerator
+// addresses from your own IP address ranges that you bring to Global Accelerator
 // (BYOIP).
 //
 // For a standard accelerator, they distribute incoming application traffic