+++ service/globalaccelerator/globalacceleratoriface/interface.go
@@ -80,6 +80,10 @@ type GlobalAcceleratorAPI interface {
 	CreateAcceleratorWithContext(aws.Context, *globalaccelerator.CreateAcceleratorInput, ...request.Option) (*globalaccelerator.CreateAcceleratorOutput, error)
 	CreateAcceleratorRequest(*globalaccelerator.CreateAcceleratorInput) (*request.Request, *globalaccelerator.CreateAcceleratorOutput)
 
+	CreateCrossAccountAttachment(*globalaccelerator.CreateCrossAccountAttachmentInput) (*globalaccelerator.CreateCrossAccountAttachmentOutput, error)
+	CreateCrossAccountAttachmentWithContext(aws.Context, *globalaccelerator.CreateCrossAccountAttachmentInput, ...request.Option) (*globalaccelerator.CreateCrossAccountAttachmentOutput, error)
+	CreateCrossAccountAttachmentRequest(*globalaccelerator.CreateCrossAccountAttachmentInput) (*request.Request, *globalaccelerator.CreateCrossAccountAttachmentOutput)
+
 	CreateCustomRoutingAccelerator(*globalaccelerator.CreateCustomRoutingAcceleratorInput) (*globalaccelerator.CreateCustomRoutingAcceleratorOutput, error)
 	CreateCustomRoutingAcceleratorWithContext(aws.Context, *globalaccelerator.CreateCustomRoutingAcceleratorInput, ...request.Option) (*globalaccelerator.CreateCustomRoutingAcceleratorOutput, error)
 	CreateCustomRoutingAcceleratorRequest(*globalaccelerator.CreateCustomRoutingAcceleratorInput) (*request.Request, *globalaccelerator.CreateCustomRoutingAcceleratorOutput)
@@ -103,6 +107,10 @@ type GlobalAcceleratorAPI interface {
 	DeleteAccelerator(*globalaccelerator.DeleteAcceleratorInput) (*globalaccelerator.DeleteAcceleratorOutput, error)
 	DeleteAcceleratorWithContext(aws.Context, *globalaccelerator.DeleteAcceleratorInput, ...request.Option) (*globalaccelerator.DeleteAcceleratorOutput, error)
 	DeleteAcceleratorRequest(*globalaccelerator.DeleteAcceleratorInput) (*request.Request, *globalaccelerator.DeleteAcceleratorOutput)
+
+	DeleteCrossAccountAttachment(*globalaccelerator.DeleteCrossAccountAttachmentInput) (*globalaccelerator.DeleteCrossAccountAttachmentOutput, error)
+	DeleteCrossAccountAttachmentWithContext(aws.Context, *globalaccelerator.DeleteCrossAccountAttachmentInput, ...request.Option) (*globalaccelerator.DeleteCrossAccountAttachmentOutput, error)
+	DeleteCrossAccountAttachmentRequest(*globalaccelerator.DeleteCrossAccountAttachmentInput) (*request.Request, *globalaccelerator.DeleteCrossAccountAttachmentOutput)
 
 	DeleteCustomRoutingAccelerator(*globalaccelerator.DeleteCustomRoutingAcceleratorInput) (*globalaccelerator.DeleteCustomRoutingAcceleratorOutput, error)
 	DeleteCustomRoutingAcceleratorWithContext(aws.Context, *globalaccelerator.DeleteCustomRoutingAcceleratorInput, ...request.Option) (*globalaccelerator.DeleteCustomRoutingAcceleratorOutput, error)
@@ -139,6 +147,10 @@ type GlobalAcceleratorAPI interface {
 	DescribeAcceleratorAttributes(*globalaccelerator.DescribeAcceleratorAttributesInput) (*globalaccelerator.DescribeAcceleratorAttributesOutput, error)
 	DescribeAcceleratorAttributesWithContext(aws.Context, *globalaccelerator.DescribeAcceleratorAttributesInput, ...request.Option) (*globalaccelerator.DescribeAcceleratorAttributesOutput, error)
 	DescribeAcceleratorAttributesRequest(*globalaccelerator.DescribeAcceleratorAttributesInput) (*request.Request, *globalaccelerator.DescribeAcceleratorAttributesOutput)
+
+	DescribeCrossAccountAttachment(*globalaccelerator.DescribeCrossAccountAttachmentInput) (*globalaccelerator.DescribeCrossAccountAttachmentOutput, error)
+	DescribeCrossAccountAttachmentWithContext(aws.Context, *globalaccelerator.DescribeCrossAccountAttachmentInput, ...request.Option) (*globalaccelerator.DescribeCrossAccountAttachmentOutput, error)
+	DescribeCrossAccountAttachmentRequest(*globalaccelerator.DescribeCrossAccountAttachmentInput) (*request.Request, *globalaccelerator.DescribeCrossAccountAttachmentOutput)
 
 	DescribeCustomRoutingAccelerator(*globalaccelerator.DescribeCustomRoutingAcceleratorInput) (*globalaccelerator.DescribeCustomRoutingAcceleratorOutput, error)
 	DescribeCustomRoutingAcceleratorWithContext(aws.Context, *globalaccelerator.DescribeCustomRoutingAcceleratorInput, ...request.Option) (*globalaccelerator.DescribeCustomRoutingAcceleratorOutput, error)
@@ -177,6 +189,24 @@ type GlobalAcceleratorAPI interface {
 
 	ListByoipCidrsPages(*globalaccelerator.ListByoipCidrsInput, func(*globalaccelerator.ListByoipCidrsOutput, bool) bool) error
 	ListByoipCidrsPagesWithContext(aws.Context, *globalaccelerator.ListByoipCidrsInput, func(*globalaccelerator.ListByoipCidrsOutput, bool) bool, ...request.Option) error
+
+	ListCrossAccountAttachments(*globalaccelerator.ListCrossAccountAttachmentsInput) (*globalaccelerator.ListCrossAccountAttachmentsOutput, error)
+	ListCrossAccountAttachmentsWithContext(aws.Context, *globalaccelerator.ListCrossAccountAttachmentsInput, ...request.Option) (*globalaccelerator.ListCrossAccountAttachmentsOutput, error)
+	ListCrossAccountAttachmentsRequest(*globalaccelerator.ListCrossAccountAttachmentsInput) (*request.Request, *globalaccelerator.ListCrossAccountAttachmentsOutput)
+
+	ListCrossAccountAttachmentsPages(*globalaccelerator.ListCrossAccountAttachmentsInput, func(*globalaccelerator.ListCrossAccountAttachmentsOutput, bool) bool) error
+	ListCrossAccountAttachmentsPagesWithContext(aws.Context, *globalaccelerator.ListCrossAccountAttachmentsInput, func(*globalaccelerator.ListCrossAccountAttachmentsOutput, bool) bool, ...request.Option) error
+
+	ListCrossAccountResourceAccounts(*globalaccelerator.ListCrossAccountResourceAccountsInput) (*globalaccelerator.ListCrossAccountResourceAccountsOutput, error)
+	ListCrossAccountResourceAccountsWithContext(aws.Context, *globalaccelerator.ListCrossAccountResourceAccountsInput, ...request.Option) (*globalaccelerator.ListCrossAccountResourceAccountsOutput, error)
+	ListCrossAccountResourceAccountsRequest(*globalaccelerator.ListCrossAccountResourceAccountsInput) (*request.Request, *globalaccelerator.ListCrossAccountResourceAccountsOutput)
+
+	ListCrossAccountResources(*globalaccelerator.ListCrossAccountResourcesInput) (*globalaccelerator.ListCrossAccountResourcesOutput, error)
+	ListCrossAccountResourcesWithContext(aws.Context, *globalaccelerator.ListCrossAccountResourcesInput, ...request.Option) (*globalaccelerator.ListCrossAccountResourcesOutput, error)
+	ListCrossAccountResourcesRequest(*globalaccelerator.ListCrossAccountResourcesInput) (*request.Request, *globalaccelerator.ListCrossAccountResourcesOutput)
+
+	ListCrossAccountResourcesPages(*globalaccelerator.ListCrossAccountResourcesInput, func(*globalaccelerator.ListCrossAccountResourcesOutput, bool) bool) error
+	ListCrossAccountResourcesPagesWithContext(aws.Context, *globalaccelerator.ListCrossAccountResourcesInput, func(*globalaccelerator.ListCrossAccountResourcesOutput, bool) bool, ...request.Option) error
 
 	ListCustomRoutingAccelerators(*globalaccelerator.ListCustomRoutingAcceleratorsInput) (*globalaccelerator.ListCustomRoutingAcceleratorsOutput, error)
 	ListCustomRoutingAcceleratorsWithContext(aws.Context, *globalaccelerator.ListCustomRoutingAcceleratorsInput, ...request.Option) (*globalaccelerator.ListCustomRoutingAcceleratorsOutput, error)
@@ -258,6 +288,10 @@ type GlobalAcceleratorAPI interface {
 	UpdateAcceleratorAttributes(*globalaccelerator.UpdateAcceleratorAttributesInput) (*globalaccelerator.UpdateAcceleratorAttributesOutput, error)
 	UpdateAcceleratorAttributesWithContext(aws.Context, *globalaccelerator.UpdateAcceleratorAttributesInput, ...request.Option) (*globalaccelerator.UpdateAcceleratorAttributesOutput, error)
 	UpdateAcceleratorAttributesRequest(*globalaccelerator.UpdateAcceleratorAttributesInput) (*request.Request, *globalaccelerator.UpdateAcceleratorAttributesOutput)
+
+	UpdateCrossAccountAttachment(*globalaccelerator.UpdateCrossAccountAttachmentInput) (*globalaccelerator.UpdateCrossAccountAttachmentOutput, error)
+	UpdateCrossAccountAttachmentWithContext(aws.Context, *globalaccelerator.UpdateCrossAccountAttachmentInput, ...request.Option) (*globalaccelerator.UpdateCrossAccountAttachmentOutput, error)
+	UpdateCrossAccountAttachmentRequest(*globalaccelerator.UpdateCrossAccountAttachmentInput) (*request.Request, *globalaccelerator.UpdateCrossAccountAttachmentOutput)
 
 	UpdateCustomRoutingAccelerator(*globalaccelerator.UpdateCustomRoutingAcceleratorInput) (*globalaccelerator.UpdateCustomRoutingAcceleratorOutput, error)
 	UpdateCustomRoutingAcceleratorWithContext(aws.Context, *globalaccelerator.UpdateCustomRoutingAcceleratorInput, ...request.Option) (*globalaccelerator.UpdateCustomRoutingAcceleratorOutput, error)